create portfolio page and change the directory structure Related to #1
